refactor: перенести MapPage в одноименнёую папку в pages
