+++ src/pages/MapPage/MapPage.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import './MapPage.css';
 
 const MapPage = () => (
   <main className='main'>
